(1)After successfully create a Banner application, return to banner application list, click on Edit button, Fix Automapper error.
(2)  Create a Save Draft button, when the Clerk clicks on it, any data filled in the fields is saved and the AppStatusID value = 4.
  After click on Submit button, AppStatusID value = 7.

(3)  Add one more dropdown field for Zone to be saved in ZoneID column in BannerObject table.

(4)After selecting multiple files to attach in required Documents, the values saved in BALinkReqDoc and Attachment table is incorrect. When testing, try uploading 2 documents for one type of Required Document.