(1)select2 BannerCode, display code number and codeDesc in one row
(2)The list of required documents have checkboxes beside it. This will not save any data in any table, it is just a condition so the "Submit" button can be clicked. The Desktop Officer must tick all checkboxes before they can submit the application to the Clerk.

(3)After clicking on Submit button, the AppStatusID will change to 2. The Desktop Officer can no longer edit this application, but they can view as read-only.